Add ability to read from JSON blobs
There are times when you may already have a JSON blob that you want
redux-json-api to work with. In those cases, there is currently no way
without using internal APIs to add it to the api state.

Example use-cases are:
1. Hydrating data via JSON blob that is passed in with the application
2. Websockets passing in data that a user wants to consume using this
   package

My proposal is to introduce `readBlob` (open to renaming - maybe
readJSON?), which takes a JSON-API object and adds it to the state.