Add ability to hydrate store
Upon application boot, you may want to hydrate the API store with
existing information. As an outside user, I should not be dependent on
the internal structure of the store for hydration. Instead, I should be
able to pass in a JSON-API blob and get the same result that I would if
I read the data from an endpoint.

In order to give the user flexibility in how to hydrate the store, as
well as preserving server-side rendering compatibility, I have opted not
to get hydration data from a global object, instead exposing the
synchronous action `hydrateStore`.

It may not make sense to call this `hydrateStore` since you could use
this for purposes other than hydration. However, `hydrateStore` seems to
fit better with the paradigms in this repository.